+++ R/regret.R
@@ -8,12 +8,20 @@
 #' @author Jacob van Etten and Kauê de Sousa
 #' @param object a data.frame, an object of class \code{pltree}, or a 
 #'  list with \code{PlackettLuce} models
-#' @param bootstrap logical, to run a bayes bootstrap on \var{object}
+#' @param bootstrap logical, to run a Bayesian bootstrap on \var{object}
 #' @param normalize logical, to normalize values to sum to 1
 #' @param group an index in \var{object} for the different scenarios 
 #' @param items an index in  \var{object} for the different items
 #' @param values an index in \var{object} with the values to compute regret
 #' @param ... further arguments passed to methods
+#' 
+#' @details
+#' Additional details for Bayesian bootstrap: 
+#' \code{statistic} A function that accepts data as its first argument and possibly, 
+#' the weights as its second, if use_weights is TRUE; \code{n1} The size of 
+#' the bootstrap sample; \code{n2} The sample size used to calculate 
+#' the statistic each bootstrap draw
+#' 
 #' @return A data frame with regret estimates
 #' \item{items}{the item names}
 #' \item{worth}{the worth parameters}
@@ -76,6 +84,7 @@
 #' mod = list(mod1, mod2)
 #' 
 #' regret(mod, n1 = 500)
+#' 
 #' @importFrom partykit nodeids
 #' @importFrom psychotools itempar
 #' @importFrom qvcalc qvcalc.itempar
@@ -125,17 +134,34 @@ regret.default = function(object, ..., values, items, group,
   # worst regret is the highest regret across the nodes
   wr = tapply(coeffs$regret, coeffs$items, max)
   
-  # regret is the sum of the squared values of all items regret
+  # max regret is the sum of the squared values of all items regret
   regret = unlist(tapply(coeffs$regret, coeffs$items, function(x) {
-    sum(x)^2
+    sum(x^2)
+  }))
+  
+  # compute standard error for regret
+  se_regret = unlist(tapply(coeffs$regret, coeffs$items, function(x) {
+    x2 = x^2
+    n = length(x2) # sample size
+    s = sd(x2) # sample standard deviation
+    s / sqrt(n)
   }))
   
   worth = tapply(coeffs$estimate, coeffs$items, mean)
   
+  # compute standard error for worth
+  se_worth = unlist(tapply(coeffs$estimate, coeffs$items, function(x) {
+    n = length(x) # sample size
+    s = sd(x) # sample standard deviation
+    s / sqrt(n)
+  }))
+  
   w = data.frame(items = items, 
                  worth = worth[items], 
                  worst_regret = wr[items], 
-                 regret = regret[items])
+                 regret = regret[items],
+                 worthSE = se_worth,
+                 regretSE = se_regret)
   
   w = w[order(w$regret), ]