+++ NEWS.md
@@ -1,3 +1,25 @@
+gosset 1.3 (2024-05-26)
+=========================
+
+### IMPROVEMENTS
+
+* Implements the function `resample()` to re-sample Plackett-Luce coefficients
+* Adds standard errors in `regret()`  
+* Adds `pairwise_probs()` to compare worth parameters following the Luce's Axiom
+
+###  BUG FIXES
+* Fix the computation of maximized regret in `regret()`. Using `sum(x^2)` instead of `sum(x)^2`
+
+
+gosset 1.2 (2023-07-03)
+=========================
+
+### IMPROVEMENTS
+
+* Implement `anova()` method for PlackettLuce models
+* Implement likelihood-ratio test for PlackettLuce models. Function `likelihood_ratio()` 
+
+
 gosset 1.0 (2023-04-18)
 =========================